Autocorrelation plot for model selection. Model inclusion of each predictor is represented across all iterations as a black and white heatmap.

Generates an autocorrelation plot from a Java MCMC results class object

AutocorrelationPlot(
  results,
  plot.title = NULL,
  file = NULL,
  include.var.names.on.x.axis = TRUE,
  covariates.to.include = NULL,
  var.dictionary = NULL,
  include.post.probs = TRUE,
  cex.x.axis.ticks = 1,
  cex.y.axis.ticks = 1,
  cex.y.axis.labels = 1,
  mar.for.varnames = 8
)

Arguments

results

A R2BGLiMS R2BGLiMS_Results-class object frum running the R2BGLiMS R2BGLiMS command.

plot.title

Optionally specify a title for the plot.

file

Path to a PDF to print plot to.

include.var.names.on.x.axis

Include variable names perpendicularly below the x-axis. Turn off if it looks too crowded.

covariates.to.include

Character vector indicating a subset of covariates to include in the plot.

var.dictionary

Character vector containing a mapping of covariate names to substitutions for the plot. The names of the elements are the variable names as in the original analysis (i.e. as named in the results object).

include.post.probs

Include a barplot of posterior probabilities at the top.

cex.x.axis.ticks

Control the size of tick marks for the covariates along the X-axis. Decrease if there are many.

cex.y.axis.ticks

Control the size of tick marks for the Y-axis.

cex.y.axis.labels

Control the size of the y-labels. Decrease if these are not fitting on.

mar.for.varnames

Margin size for the bottom axis. Default is 8 so there is room for perpendicular variable names. Decrease if not using variable names.

Value

NA

Author

Paul Newcombe