Logistic to linear effect conversion for JAM

Given marginal log-ORs and corresponding MAFs, this function infers the effect estimates that would have been obtained running one-at-a-time linear regressions of the binary outcome. This conversion is required for JAM since it is based on a linear regression model.

JAM_LogisticToLinearEffects(
  log.ors = NULL,
  log.or.ses = NULL,
  snp.genotype.sds = NULL,
  mafs = NULL,
  n = NULL,
  p.cases = NULL
)

Arguments

log.ors

Vector of SNP summary log-odds ratios

log.or.ses

Vector of the standard errors of the SNP log-odds ratios

snp.genotype.sds

Vector of SNP genotype standard deviations. Please supply this or mafs. This is the preferred option since an assumption of Hardy-Weinberg Equilibrium is not required when reverse-stamdardising the effects during the transform.

mafs

Vector of SNP minor allele frequencies. snp.sds if available would be the preferred option.

n

Size of dataset in which the log-odds ratios were calculated

p.cases

Proportion of cases in the dataset in which the log-odds ratios were calculated

Value

A vector of effects on the linear scale

Author

Paul Newcombe