A pretty summary results table for reports. NOTE: This function outputs a table formatted with character strings. A numeric representation of the results are stored in the slot 'posterior.summary.table'.

Creates a `pretty' summary results table from a Reversible Jump results object

PrettyResultsTable(
  results,
  round.digits.betas = 2,
  round.digits.postprob = 2,
  round.digits.bf = 1,
  normalised.sds = NULL
)

Arguments

results

A R2BGLiMS R2BGLiMS_Results-class object frum running the R2BGLiMS R2BGLiMS command.

round.digits.betas

Number of decimal places to include for effect estimates. (Default is 2)

round.digits.postprob

Number of decimal places to include for posterior probabilities. (Default is 2)

round.digits.bf

Number of decimal places to include for Bayes factors. (Default is 1)

normalised.sds

If covariates were normalised by their standard deviation, provide a named vector of the standard deviations on the original scale using this argument. Effects in the resulting table will then be interpretable according to unit changes on each covariate's original scale, rather than according to standard deviation changes.

Value

A nice summary results table.

Author

Paul Newcombe