Parameter posterior trace plots for a R2BGLiMS results object

Generates posterior trace plots for a R2BGLiMS results object

TracePlots(
  results,
  vars.to.include = NULL,
  var.dictionary = NULL,
  predictors.only = FALSE,
  file = NULL,
  add.to.title = NULL,
  par.mfrow = NULL,
  add.panel.labels = FALSE
)

Arguments

results

A R2BGLiMS R2BGLiMS_Results-class object frum running the R2BGLiMS R2BGLiMS command.

vars.to.include

Check NULL values.

var.dictionary

Character vector containing a mapping of covariate names to substitutions for the plot. The names of the elements are the variable names as in the original analysis (i.e. as named in the results object).

predictors.only

If true only chain plots for predictors are plotted (and intercept, likelihood etc. are excluded). Default is FALSE.

file

PDF path

add.to.title

Optional character string to add to the start of each plot title

par.mfrow

Allow passing the mfrow layout vector to par. Default par(mfrow=c(5,2))

add.panel.labels

Add lower case letter labels to panel titles. Default FALSE. Note: Does not work when providing a var.dictionary

Value

NA

Author

Paul Newcombe