Wrapper to run FINEMAP (Benner et al. 2016) in R

Wrapper to run FINEMAP (Benner et al. 2016) in R

finemap(GWAS, ldfile, N, maxcv = 10, fstub, FMpath)

Arguments

GWAS

a data.frame with columns: "rsid", "chromosome", "position", "allele1", "allele2", "maf", "beta", "se"; this is the same as z file of FINEMAP

ldfile

path to file that contains the SNP correlation matrix, with SNPs in the same order as in the GWAS data.frame; this is the same as the ld file of FINEMAP

N

sample size for trait

maxcv

maximum number of causal variants

fstub

file stub for input/output files of FINEMAP, e.g. if fstub="DIRresults/region1", FINEMAP files of the form "DIRresults/region1.z" will be created

FMpath

file pathway to FINEMAP software e.g. "/software/finemap_v1.4_x86_64/finemap_v1.4_x86_64"

Value

snpPP a data.frame of top SNP models and their model PPs, as output from FINEMAP

Author

Jenn Asimit